src/app/components/home/timeline-item/timeline-item.component.ts
import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
import { CommonModule } from '@angular/common';

@Component({
  selector: 'app-timeline-item',
  templateUrl: './timeline-item.component.html',
  styleUrls: ['./timeline-item.component.scss'],
  standalone: true,
  changeDetection: ChangeDetectionStrategy.OnPush,
  imports: [CommonModule],
})
export class TimelineItemComponent {
  @Input({ required: true }) chip = ''
  @Input({ required: true }) date = ''
  @Input({ required: true }) header = ''
}
